Unify get_top_of_ram and get_cbmem_toc

+++ src/northbridge/amd/lx/northbridgeinit.c
@@ -30,6 +30,7 @@
 #include <cpu/amd/lxdef.h>
 #include <cpu/x86/msr.h>
 #include <cpu/x86/cache.h>
+#include <cbmem.h>
 
 struct gliutable {
 	unsigned long desc_name;
@@ -713,7 +714,7 @@ static void setup_lx_cache(void)
 	wbinvd();
 }
 
-uint32_t get_top_of_ram(void)
+unsigned long get_top_of_ram(void)
 {
 	struct gliutable *gl = 0;
 	uint32_t systop;
